requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
appdirs==1.4.4
astroid==2.3.3
autopep8==1.5
colorama==0.4.3
distlib==0.3.0
django-cors-headers==2.1.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2020.6.2
pycodestyle==2.5.0
pylint==2.4.4
virtualenv==20.0.23
virtualenv-clone==0.5.4
wrapt==1.11.2